requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
pytest==7.4.4
selenium==4.17.2
